requirements.txt
streamlit
langchain
langchain-openai
langchain-google-genai
chromadb
pypdf
python-dotenv
tiktoken
sentence-transformers
